+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-# Improvements
+# New Features and Improvements
 
-- Fixed the wrong logic of `chatMaxLengthBehavior` of the previous version.
-- Added `sendByParts` option for `chatMaxLengthBehavior`, and two extra toggles for it: `sendByPartsPartSize` and `sendByPartsIntervalMillis`.
+- Added `/mserver resourcepacks` command to view paths of downloaded server resourcepacks